refactor(geometry): move apply_submesh_placement into transforms
The module_size_ratchet allowlists router/processing.rs at a frozen
1183-line budget and the unplaced-frame capture pushed it to 1190.
Move the whole placement applier next to apply_placement in
transforms/mod.rs (placement logic lives there anyway): processing.rs
drops to 1141, transforms/mod.rs stays under the 400-line limit at 328.
No behavior change.
